Arrangement

+++ src/main/java/rentcar/backend/business/concretes/BrandManager.java
@@ -3,6 +3,7 @@ package rentcar.backend.business.concretes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rentcar.backend.business.abstracts.BrandService;
+import rentcar.backend.core.exception.AlreadyExistsException;
 import rentcar.backend.core.exception.NotFoundException;
 import rentcar.backend.dataaccess.abstracts.BrandRepository;
 import rentcar.backend.entities.concrete.Brand;
@@ -14,6 +15,8 @@ import java.util.Optional;
 public class BrandManager implements BrandService {
 
     private final BrandRepository brandRepository;
+
+    private Optional<Brand> optionalBrand;
 
     @Autowired
     public BrandManager(BrandRepository brandRepository){
@@ -27,12 +30,19 @@ public class BrandManager implements BrandService {
 
     @Override
     public Brand addBrand(Brand brand) {
+        if (!brandIsPresent(brand)){
+            throw new AlreadyExistsException("This brand already exist!");
+        }
         return brandRepository.save(brand);
     }
 
     @Override
     @Transactional
     public void deleteBrand(int id) {
+        optionalBrand=brandRepository.findById(id);
+        if (optionalBrand.isPresent()){
+            throw new NotFoundException("Brand is not found!");
+        }
         brandRepository.deleteById(id);
     }
 
@@ -46,7 +56,7 @@ public class BrandManager implements BrandService {
     }
 
     private boolean brandIsPresent(Brand brand){
-        Optional<Brand> optionalBrand= brandRepository.findById(brand.getId());
+        optionalBrand= brandRepository.findById(brand.getId());
         if (!optionalBrand.isPresent()){
             return true;
         }
